real_aura/requirements.txt
# Core dependencies - no fluff, just what we need
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
psutil==5.9.6
requests==2.31.0
websocket-client==1.6.4
rich==13.7.0
prometheus-client==0.19.0
pydantic==2.5.0

# For development
pytest==7.4.3
httpx==0.25.2